fix: extract showToast to shared module, fix ESC listener leak in modals
- Move showToast() into core/uiHelpers.ts to eliminate duplication
- Add attachModalDismiss() helper for consistent overlay dismiss behaviour
- Fix ESC keydown listener not being cleaned up when overlay is dismissed by clicking outside

Made-with: Cursor
