+++ src/main.ts
@@ -16,6 +16,7 @@ import {
 } from './core/templateStore';
 import { testNowPlayingConnection } from './core/nowPlayingProvider';
 import { initCopyUrlButton } from './core/copyUrl';
+import { showToast, attachModalDismiss } from './core/uiHelpers';
 
 console.log('%cPV Tool%c solaris:0914', 'color:#6688cc;font-weight:bold', 'color:#888');
 
@@ -43,18 +44,9 @@ function showModal(contentHtml: string, confirmText: string): void {
   `;
   document.body.appendChild(overlay);
 
-  const confirmBtn = overlay.querySelector('.pv-modal-confirm')!;
-  confirmBtn.addEventListener('click', () => overlay.remove());
-  overlay.addEventListener('click', (e) => {
-    if (e.target === overlay) overlay.remove();
-  });
-  const onEsc = (e: KeyboardEvent) => {
-    if (e.key === 'Escape') {
-      overlay.remove();
-      document.removeEventListener('keydown', onEsc);
-    }
-  };
-  document.addEventListener('keydown', onEsc);
+  overlay.querySelector('.pv-modal-confirm')!
+    .addEventListener('click', () => overlay.remove());
+  attachModalDismiss(overlay);
 }
 
 app.innerHTML = `
@@ -601,15 +593,6 @@ tplDeleteOk.addEventListener('click', () => {
   syncSpeedSlider();
   tplDeleteConfirm.style.display = 'none';
 });
-
-// Toast helper
-function showToast(msg: string) {
-  const el = document.createElement('div');
-  el.className = 'pv-toast';
-  el.textContent = msg;
-  document.body.appendChild(el);
-  setTimeout(() => el.remove(), 2200);
-}
 
 // Export share code
 tplExportBtn.addEventListener('click', async () => {
